move of mapper to ruby event store
