+++ rails_event_store_active_record/lib/rails_event_store_active_record.rb
@@ -1,8 +1,6 @@
 require 'rails_event_store_active_record/generators/migration_generator'
 require 'rails_event_store_active_record/generators/v1_v2_migration_generator'
 require 'rails_event_store_active_record/event'
-require 'rails_event_store_active_record/mappers/yaml'
-require 'rails_event_store_active_record/mappers/with_event_class_remapping'
 require 'rails_event_store_active_record/event_repository'
 require 'rails_event_store_active_record/legacy_event_repository'
 require 'rails_event_store_active_record/version'
