+++ ruby_event_store/lib/ruby_event_store.rb
@@ -8,4 +8,5 @@ require 'ruby_event_store/client'
 require 'ruby_event_store/event'
 require 'ruby_event_store/deprecations'
 require 'ruby_event_store/serialized_record'
+require 'ruby_event_store/mappers/default'
 require 'ruby_event_store/version'
